deprecated load_tree : first try
